[design] 팀탭 퍼블리싱, todo 컴포넌트로 뺌
