+++ src/components/(suwan)/my-study/my-todo/MyTodoContainer.tsx
@@ -2,9 +2,9 @@ import Calender from '@/components/(suwan)/my-study/common/Calender';
 import Todos from '@/components/(suwan)/my-study/my-todo/Todos';
 export default function MyTodoContainer() {
   return (
-    <>
+    <div className="bg-gray-100">
       <Calender />
       <Todos />
-    </>
+    </div>
   );
 }
